feat: add add-commit-hash.js
