+++ packages/chatbot-ui/package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "dev": "DEV_MODE=true dotenv vite",
     "build": "vite build --mode client && vite build --mode server",
-    "build:lib": "vite build --mode lib && tsc --project tsconfig.lib.json",
+    "build:lib": "vite build --mode lib && tsc --project tsconfig.lib.json && node scripts/add-commit-hash.js",
     "start": "node dist/index.js",
     "lint": "tsc && biome lint --error-on-warnings ./src",
     "format": "biome check --write --unsafe .",
